fixed distanceTo geofence

+++ app/src/main/java/charlyn23/c4q/nyc/projectx/Constants.java
@@ -28,7 +28,7 @@ public class Constants {
     public static final String CREATED_AT = "createdAt";
     public static final long MILLI_24HOURS = 86400000;
     public static final long MILLI_48HOURS = 172800000;
-    public static final double FIFTY_METERS = 80467.2;
+    public static final float FIFTY_METERS = 80467.2f;
     public static final float GEOFENCE_RADIUS = 1609; // 1 mile, 1.6 km
     public static final long GEOFENCE_EXPIRATION_IN_HOURS = 24;
     public static final long GEOFENCE_EXPIRATION_IN_MILLISECONDS = GEOFENCE_EXPIRATION_IN_HOURS * 60 * 60 * 1000;
